Split application to angular modules there are two modules currently 'miqStaticAssets.toolbar' and 'miqStaticAssets.common', these angular modules are included in main application 'miqStaticAssets' so basic modules does not have to be included one by one
