+++ src/common/index.ts
@@ -1,0 +1,7 @@
+///<reference path="../tsd.d.ts"/>
+import services from './services';
+
+module common {
+  export const app = angular.module('miqStaticAssets.common', []);
+  services(app);
+}
