+++ src/common/interfaces/endpoints.ts
@@ -1,0 +1,8 @@
+export interface IEndpoints {
+  listDataTable: string;
+  deleteItemDataTable: string;
+  validateItem: string;
+  createItem: string;
+  providerSettings: string;
+  toolbarSettings: string;
+}
