+++ src/common/services/index.ts
@@ -1,0 +1,5 @@
+import EndpointsService from './endpointsService';
+
+export default (module: ng.IModule) => {
+  module.service('MiQEndpointsService', EndpointsService);
+};
