+++ src/toolbar/services/toolbarSettingsService.ts
